pyproject.toml
[project]
name = "edusmart-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "MKhai205",email = "manhkhai123459876@gmail.com"}
]
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "google-auth (>=2.41.1,<3.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "langchain (>=1.2.12,<2.0.0)",
    "langchain-google-genai (>=4.2.1,<5.0.0)",
    "unstructured[pdf] (>=0.21.5,<0.22.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "alembic (>=1.16.5,<2.0.0)",
    "ruff (>=0.15.6,<0.16.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
